Re-purposing the task flag migration in issue_statuses table as a default flag.
Unlikely anyone will run the old migration (except me) :)

We want to flag default statuses (a subset of those in the workflows
for this tracker) that will be used by default for all projects.
Projects can then add more in as they wish.

The motivation here is that there is a preference to stick to simpler
taskboards and only added extra columns if the project has special
requirements.